+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Packing to Docker image
 - Check filemodes before packing
 - `--from` option for `docker pack` command to specify base image Dockerfile path
+- `cartridge.pre-build` and `cartridge.post-build` hooks
+  to be ran before and after `rocks make`
 
 ### Changed